Revert "split out common E2E tests"
This reverts commit be8f4f8461a56bc32dd27aa3b8d7a2ae8d8bf063.
